- Improvements to Spike sorter - Must put public namespaces in their own dll
git-svn-id: https://neurorighter.googlecode.com/svn/trunk@383 471701b0-e017-11dd-9759-139c9ab14fc6

+++ simoc/spk2obs/Spk2ASDR.cs
@@ -25,7 +25,7 @@ namespace simoc.spk2obs
         internal override void MeasureObservable()
         {
             // Estimate the CSDR
-            currentObservation = (double)newSpikes.eventBuffer.Count() / numSecondInCurrentRead; 
+            currentObservation = (double)newSpikes.eventBuffer.Count() / (double)numSecondInCurrentRead; 
         }
